敵が環境の影響を受けるようにしたい！
