+++ Assets/山本/山本_Script/CharacterBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 821f7e9444f97374792293e96f522d82
